+++ Arma3ObjectBuilder/io_p3d/utils.py
@@ -1,0 +1,13 @@
+def clear_uvs(obj):
+    uvs = [uv for uv in obj.data.uv_layers]
+
+    while uvs:
+        obj.data.uv_layers.remove(uvs.pop())
+
+
+def create_selection(obj, selection):
+    group = obj.vertex_groups.get(selection, None)
+    if not group:
+        group = obj.vertex_groups.new(name=selection.strip())
+
+    group.add([vert.index for vert in obj.data.vertices], 1, 'REPLACE')
