+++ Arma3ObjectBuilder/props/material.py
@@ -3,7 +3,8 @@ import re
 import bpy
 
 from .. import get_prefs
-from ..utilities import generic as utils
+from .. import utils
+from .. import utils_io
 from ..utilities import data
 
 
@@ -75,9 +76,9 @@ class A3OB_PG_properties_material(bpy.types.PropertyGroup):
                 self.color_raw = texture
             
         else:
-            self.texture_path = utils.restore_absolute(texture) if absolute else texture
+            self.texture_path = utils_io.restore_absolute(texture) if absolute else texture
         
-        self.material_path = utils.restore_absolute(material) if absolute else material
+        self.material_path = utils_io.restore_absolute(material) if absolute else material
     
     def to_p3d(self, relative):
         addon_prefs = get_prefs()
@@ -85,14 +86,14 @@ class A3OB_PG_properties_material(bpy.types.PropertyGroup):
         material = ""
 
         if self.texture_type == 'TEX':
-            texture = utils.format_path(utils.abspath(self.texture_path), utils.abspath(addon_prefs.project_root), relative)
+            texture = utils_io.format_path(utils_io.abspath(self.texture_path), utils_io.abspath(addon_prefs.project_root), relative)
         elif self.texture_type == 'COLOR':
             color = self.color_value
             texture = "#(argb,8,8,3)color(%.3f,%.3f,%.3f,%.3f,%s)" % (color[0], color[1], color[2], color[3], self.color_type)
         else:
             texture = self.color_raw
         
-        material = utils.format_path(utils.abspath(self.material_path), utils.abspath(addon_prefs.project_root), relative)
+        material = utils_io.format_path(utils_io.abspath(self.material_path), utils_io.abspath(addon_prefs.project_root), relative)
 
         return texture, material
