+++ Arma3ObjectBuilder/ui/tool_hitpoint.py
@@ -1,7 +1,7 @@
 import bpy
 
 from .. import get_icon
-from ..utilities import generic as utils
+from .. import utils
 from ..utilities import clouds as cloudutils
