+++ Arma3ObjectBuilder/ui/tool_mass.py
@@ -3,7 +3,7 @@ from math import log10, ceil
 import bpy
 
 from .. import get_icon
-from ..utilities import generic as utils
+from .. import utils
 from ..utilities import masses as massutils
