+++ Arma3ObjectBuilder/utils.py
@@ -4,14 +4,13 @@
 import os
 import json
 from contextlib import contextmanager
-from datetime import datetime
 
 import bpy
 import bpy_extras.mesh_utils as meshutils
 import bmesh
 
-from .. import get_prefs
-from . import data
+from . import get_prefs
+from .utilities import data
 
 
 # For some reason, not all operator reports are printed to the console. The behavior seems to be context dependent,
@@ -19,13 +18,6 @@ from . import data
 def op_report(operator, mode, message):
     operator.report(mode, message)
     print("%s: %s\n" % (tuple(mode)[0].title(), message))
-
-
-def abspath(path):
-    if not path.startswith("//"):
-        return path
-    
-    return os.path.abspath(bpy.path.abspath(path))
 
 
 def is_valid_idx(index, subscriptable):
@@ -151,73 +143,8 @@ def force_mode_edit():
     bpy.ops.object.mode_set(mode='EDIT')
 
 
-def create_selection(obj, selection):
-    group = obj.vertex_groups.get(selection, None)
-    if not group:
-        group = obj.vertex_groups.new(name=selection.strip())
-
-    group.add([vert.index for vert in obj.data.vertices], 1, 'REPLACE')
-
-
-def clear_uvs(obj):
-    uvs = [uv for uv in obj.data.uv_layers]
-
-    while uvs:
-        obj.data.uv_layers.remove(uvs.pop())
-
-
-def replace_slashes(path):
-    return path.replace("/", "\\")
-
-
-# Attempt to restore absolute paths to the set project root (P drive by default).
-def restore_absolute(path, extension = ""):
-    path = replace_slashes(path.strip().lower())
-    
-    if path == "":
-        return ""
-    
-    if os.path.splitext(path)[1].lower() != extension:
-        path += extension
-    
-    root = abspath(get_prefs().project_root).lower()
-    if not path.startswith(root):
-        abs_path = os.path.join(root, path)
-        if os.path.exists(abs_path):
-            return abs_path
-    
-    return path
-
-
-def make_relative(path, root):
-    path = path.lower()
-    root = root.lower()
-    
-    if root != "" and path.startswith(root):
-        return os.path.relpath(path, root)
-    
-    drive = os.path.splitdrive(path)[0]
-    if drive:
-       path = os.path.relpath(path, drive)
-    
-    return path
-
-
-def format_path(path, root = "", to_relative = True, extension = True):
-    path = replace_slashes(path.strip())
-    
-    if to_relative:
-        root = replace_slashes(root.strip())
-        path = make_relative(path, root)
-        
-    if not extension:
-        path = os.path.splitext(path)[0]
-        
-    return path
-
-
 def load_common_data(scene):
-    custom_path = abspath(get_prefs().custom_data)
+    custom_path = bpy.path.abspath(get_prefs().custom_data)
     builtin = data.common_data
     json_data = {}
     try:
@@ -238,40 +165,3 @@ def load_common_data(scene):
             item.type = category.upper()
     
     scene_props.items_index = 0
-
-
-class ExportFileHandler():
-    def __init__(self, filepath, mode):
-        timestamp = datetime.now().strftime("%Y%m%d%H%M%S")
-        self.filepath = filepath
-        self.temppath = "%s.%s.temp" % (filepath, timestamp)
-        self.mode = mode
-        self.file = None
-        addon_pref = get_prefs()
-        self.backup_old = addon_pref.create_backups
-        self.preserve_faulty = addon_pref.preserve_faulty_output
-
-    def __enter__(self):
-        file = open(self.temppath, self.mode)
-        self.file = file
-
-        return file
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self.file.close()
-
-        if exc_type is None:
-            if os.path.isfile(self.filepath) and self.backup_old:
-                self.force_rename(self.filepath, self.filepath + ".bak")
-
-            self.force_rename(self.temppath, self.filepath)
-        
-        elif not self.preserve_faulty:
-            os.remove(self.temppath)
-    
-    @staticmethod
-    def force_rename(old, new):
-        if os.path.isfile(new):
-            os.remove(new)
-        
-        os.rename(old, new)
